+++ mss/src/core/utility/flat_hashtable.h
@@ -13,7 +13,7 @@ struct SplitMix64Hash {
         x += 0x9e3779b97f4a7c15ULL;
         x = (x ^ (x >> 30)) * 0xbf58476d1ce4e5b9ULL;
         x = (x ^ (x >> 27)) * 0x94d049bb133111ebULL;
-        return static_cast<std::size_t>(x ^ (x >> 31));
+        return x ^ (x >> 31);
     }
 };
